✨ Integrate Role struct of werewolf_cli into src/werewolf/type
