+++ examples/werewolf_cli/game/player.rs
@@ -1,18 +1,19 @@
 use crate::game::role::Role;
 use serde::{Deserialize, Serialize};
+use zk_mpc::werewolf::types::Role;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Player {
     pub id: usize,
     pub name: String,
-    pub role: Role,
+    pub role: Option<Role>,
     pub is_alive: bool,
     pub death_day: Option<u32>,
     pub marked_for_death: bool,
 }
 
 impl Player {
-    pub fn new(id: usize, name: String, role: Role) -> Self {
+    pub fn new(id: usize, name: String, role: Option<Role>) -> Self {
         Self {
             id,
             name,
@@ -29,7 +30,7 @@ impl Player {
     }
 
     pub fn is_werewolf(&self) -> bool {
-        self.role.is_werewolf()
+        self.role.unwrap().is_werewolf()
     }
 
     pub fn mark_for_death(&mut self) {
@@ -37,7 +38,12 @@ impl Player {
     }
 }
 
-pub fn create_players(names: Vec<String>, roles: Vec<Role>) -> Vec<Player> {
+pub fn create_players(names: Vec<String>, roles: Option<Vec<Role>>) -> Vec<Player> {
+    let roles = match roles {
+        Some(roles) => roles.into_iter().map(Some).collect(),
+        None => vec![None; names.len()],
+    };
+
     names
         .into_iter()
         .zip(roles)
