+++ examples/werewolf_cli/game/role.rs
@@ -1,29 +1,9 @@
 use rand::seq::SliceRandom;
 use rand::thread_rng;
 use serde::{Deserialize, Serialize};
+use zk_mpc::werewolf::types::Role;
 
 use super::GameRules;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum Role {
-    Villager,
-    Werewolf,
-    Seer,
-}
-
-impl Role {
-    pub fn description(&self) -> &'static str {
-        match self {
-            Role::Villager => "村人：特別な能力はありませんが、議論と投票に参加します。",
-            Role::Werewolf => "人狼：夜に村人を襲撃します。昼は村人のふりをします。",
-            Role::Seer => "占い師：夜に一人のプレイヤーの役割を知ることができます。",
-        }
-    }
-
-    pub fn is_werewolf(&self) -> bool {
-        matches!(self, Role::Werewolf)
-    }
-}
 
 pub fn assign_roles(player_count: usize, rules: &GameRules) -> Vec<Role> {
     let mut roles = vec![Role::Villager; player_count];
@@ -38,7 +18,7 @@ pub fn assign_roles(player_count: usize, rules: &GameRules) -> Vec<Role> {
     }
 
     for role in roles.iter_mut().skip(werewolf_count).take(seer_count) {
-        *role = Role::Seer;
+        *role = Role::FortuneTeller;
     }
 
     // ランダムに並び替える
